refactor: move interface to its own directory. context folder changed to hooks.
