+++ src/pages/Dashboard/UserAdList/UserAdList.tsx
@@ -1,11 +1,11 @@
 import { useEffect, useState } from 'react'
-import { AdType } from '../../SearchAds/SearchAds'
 import UserAdListItem from './UserAdListItem/UserAdListItem'
 import { Query } from 'appwrite'
 import db from '../../../appwrite/databases'
+import { IAd } from '../../../types'
 
 const UserAdList = () => {
-    const [ads, setAds] = useState<AdType[]>([])
+    const [ads, setAds] = useState<IAd[]>([])
 
     useEffect(() => {
         init();
@@ -23,7 +23,7 @@ const UserAdList = () => {
             <div>UserAdList</div>
             <div>
                 {ads.map(
-                    (ad: AdType) => (
+                    (ad: IAd) => (
                         <UserAdListItem key={ad.$id} setAds={setAds} adData={ad} />
                     )
                 )}
