+++ src/types/index.ts
@@ -1,0 +1,6 @@
+import { Models } from "appwrite";
+
+export interface IAd extends Models.Document {
+    title: string,
+    body: string
+}
