+++ src/utils/ProtectedRoutes.tsx
@@ -1,5 +1,5 @@
 import { Outlet, Navigate } from 'react-router-dom'
-import { useAuth } from '../contexts/AuthContext';
+import { useAuth } from '../hooks/AuthContext';
 
 const ProtectedRoutes = () => {
     const auth = useAuth();
